Abstract Shabal and Curve25519 functions (#74)
* Make Shabal256 conform to MessageDigest

* Remove ShabalGeneric

* Abstract Curve25519

* Test, Bug Fixes

* Tweak

* Silly.

* Minor performance tweaks
