+++ src/brs/Burst.java
@@ -4,6 +4,7 @@ import brs.AT.HandleATBlockTransactionsListener;
 import brs.assetexchange.AssetExchange;
 import brs.assetexchange.AssetExchangeImpl;
 import brs.blockchainlistener.DevNullListener;
+import brs.crypto.hash.ShabalProvider;
 import brs.db.BlockDb;
 import brs.db.cache.DBCacheManagerImpl;
 import brs.db.sql.Db;
@@ -156,6 +157,8 @@ public final class Burst {
 
     try {
       long startTime = System.currentTimeMillis();
+
+      ShabalProvider.init();
 
       final TimeService timeService = new TimeServiceImpl();
